+++ packages/webcomponents/src/actions/terminal/get-terminal-detail.ts
