+++ packages/webcomponents/src/actions/terminal/get-terminals.ts
