+++ packages/webcomponents/src/components/business-details/test/busines-details-core.spec.tsx
@@ -3,7 +3,7 @@ jest.mock('../../../ui-components/styled-host/styled-host.css', () => '');
 import { h } from "@stencil/core";
 import { newSpecPage } from "@stencil/core/testing";
 import { BusinessDetailsCore } from "../business-details-core";
-import { makeGetBusiness } from "../get-business";
+import { makeGetBusiness } from "../../../actions/business/get-business";
 import mockBusinessDetails from './../../../../../../mockData/mockBusinessDetails.json';
 import { Business, IBusiness } from "../../../api/Business";
 import { API_NOT_AUTHENTICATED_ERROR } from "../../../api/shared";
