+++ packages/webcomponents/src/components/business-details/test/get-business.spec.ts
@@ -1,4 +1,4 @@
-import { makeGetBusiness } from '../get-business';
+import { makeGetBusiness } from '../../../actions/business/get-business';
 import mockResponse from '../../../../../../mockData/mockBusinessDetails.json';
 import { Business, IBusiness } from '../../../api/Business';
 import { API_NOT_AUTHENTICATED_ERROR } from '../../../api/shared';
