+++ packages/webcomponents/src/components/gross-payment-chart/test/get-gross-payment-chart-data.spec.ts
@@ -1,4 +1,4 @@
-import { makeGetGrossPaymentChartData } from '../get-gross-payment-chart-data';
+import { makeGetGrossPaymentChartData } from '../../../actions/gross-payment/get-gross-payment-chart-data';
 import { ReportsService } from '../../../api/services/reports.service';
 import { mockGrossVolumeReport } from '../../../../../../mockData/mockGrossVolumeReport';
 import { API_NOT_AUTHENTICATED_ERROR } from '../../../api/shared';
