+++ packages/webcomponents/src/components/payment-details/test/payment-details-core.spec.tsx
@@ -5,7 +5,7 @@ import { newSpecPage } from '@stencil/core/testing';
 import mockPaymentDetailsResponse from '../../../../../../mockData/mockPaymentDetailSuccess.json';
 import { PaymentDetailsCore } from '../payment-details-core';
 import { Details } from '../../../ui-components/details/details';
-import { makeGetPaymentDetails } from '../get-payment-details';
+import { makeGetPaymentDetails } from '../../../actions/payment/get-payment-details';
 import { API_NOT_AUTHENTICATED_ERROR } from '../../../api/shared';
 
 describe('payment-details-core', () => {
