+++ packages/webcomponents/src/components/payout-details/test/get-payout-details.spec.ts
@@ -1,6 +1,6 @@
 import { IApiResponse, Payout } from '../../../api';
 import { PayoutService } from '../../../api/services/payout.service';
-import { makeGetPayoutDetails } from '../get-payout-details';
+import { makeGetPayoutDetails } from '../../../actions/payout/get-payout-details';
 import mockResponse from '../../../../../../mockData/mockPayoutDetailsSuccess.json';
 
 describe('getPayout', () => {
