+++ packages/webcomponents/src/components/payout-details/test/payout-details-core.spec.tsx
@@ -4,7 +4,7 @@ import { h } from '@stencil/core';
 import { newSpecPage } from '@stencil/core/testing';
 import { PayoutDetailsCore } from '../payout-details-core';
 import mockPayoutDetailSuccess from '../../../../../../mockData/mockPayoutDetailsSuccess.json';
-import { makeGetPayoutDetails } from '../get-payout-details';
+import { makeGetPayoutDetails } from '../../../actions/payout/get-payout-details';
 import { API_NOT_AUTHENTICATED_ERROR } from '../../../api/shared';
 
 describe('payout-details-core', () => {
